requirements.txt
fastapi
uvicorn[standard]
watchdog
sqlmodel
pydantic-settings
plyer
python-dotenv
pyyaml
httpx
pathspec
